# src/handler/auth_handler.go
package handler

import (
	"hi-go/src/model"
	"hi-go/src/service"
	"strconv"

	"github.com/gin-gonic/gin"
)

// 认证处理器
type AuthHandler struct {
	authService *service.AuthService
}

// 创建认证处理器实例
func NewAuthHandler() *AuthHandler {
	return &AuthHandler{
		authService: service.NewAuthService(),
	}
}

// Login 登录接口
// @Summary      用户登录
// @Description  用户通过用户名和密码登录，返回用户信息和 JWT Token
// @Tags         认证模块
// @Accept       json
// @Produce      json
// @Param        request  body      model.LoginRequest   true  "登录请求参数"
// @Success      200      {object}  model.Response{data=model.LoginData}  "登录成功"
// @Failure      400      {object}  model.Response  "参数错误"
// @Failure      401      {object}  model.Response  "认证失败"
// @Router       /user/login [post]
func (h *AuthHandler) Login(c *gin.Context) {
	var req model.LoginRequest

	// 1. 绑定并验证请求参数
	if err := c.ShouldBindJSON(&req); err != nil {
		model.ParamError(c, "参数错误: "+err.Error())
		return
	}

	// 2. 调用服务层登录
	resp, err := h.authService.Login(&req)
	if err != nil {
		model.Unauthorized(c, err.Error())
		return
	}

	// 3. 返回成功响应
	model.Success(c, resp)
}

// Register 注册接口
// @Summary      用户注册
// @Description  创建新用户账号
// @Tags         认证模块
// @Accept       json
// @Produce      json
// @Param        request  body      model.RegisterRequest  true  "注册请求参数"
// @Success      200      {object}  model.Response{data=model.User}  "注册成功"
// @Failure      400      {object}  model.Response  "参数错误"
// @Router       /user/register [post]
func (h *AuthHandler) Register(c *gin.Context) {
	var req model.RegisterRequest

	// 1. 绑定并验证请求参数
	if err := c.ShouldBindJSON(&req); err != nil {
		model.ParamError(c, "参数错误: "+err.Error())
		return
	}

	// 2. 调用服务层注册
	user, err := h.authService.Register(&req)
	if err != nil {
		model.ParamError(c, err.Error())
		return
	}

	// 3. 返回成功响应
	model.SuccessWithMessage(c, "注册成功", user)
}

// GetProfile 获取用户信息
// @Summary      获取当前用户信息
// @Description  获取当前登录用户的详细信息
// @Tags         认证模块
// @Accept       json
// @Produce      json
// @Security     BearerAuth
// @Success      200  {object}  model.Response{data=model.User}  "获取成功"
// @Failure      401  {object}  model.Response  "未授权"
// @Failure      404  {object}  model.Response  "用户不存在"
// @Router       /user/profile [get]
func (h *AuthHandler) GetProfile(c *gin.Context) {
	// 从上下文获取用户ID（中间件已设置）
	userID, exists := c.Get("userID")
	if !exists {
		model.Unauthorized(c, "未授权")
		return
	}

	// 将 string 类型的 userID 转换为 int64
	userIDStr, ok := userID.(string)
	if !ok {
		model.ParamError(c, "用户ID格式错误")
		return
	}

	userIDInt64, err := strconv.ParseInt(userIDStr, 10, 64)
	if err != nil {
		model.ParamError(c, "用户ID无效")
		return
	}

	// 获取用户信息
	user, err := h.authService.GetUserByID(userIDInt64)
	if err != nil {
		model.NotFound(c, err.Error())
		return
	}

	model.Success(c, user)
}